Move proudction env files to sub folder
