+++ src/testProductionEnvironments/environmentsByUrl.ts
@@ -1,10 +1,5 @@
-import { eb } from './eb';
+import { eb } from '../eb';
 import { envNames } from './environments';
-import { setIsHealthy } from './health';
-
-process.on('unhandledRejection', () => {
-  setIsHealthy(false);
-});
 
 export const environmentsByUrl = eb
   .describeEnvironments({
